refactor: add rise fall description
