+++ packages/trader/src/Assets/Trading/Categories/Description/accumulator-trade-description.tsx
