+++ packages/trader/src/Assets/Trading/Categories/Description/turbos-trade-description.tsx
