+++ packages/trader/src/Assets/Trading/Categories/__tests__/turbos-trade-description.spec.tsx
@@ -1,6 +1,6 @@
 import React from 'react';
 import { render, screen } from '@testing-library/react';
-import { TurbosTradeDescription } from '../turbos-trade-description';
+import { TurbosTradeDescription } from '../Description/turbos-trade-description';
 
 describe('<TurbosTradeDescription />', () => {
     it('a proper text of description should be rendered', () => {
